Add unimplemented deleteMax() and max() methods with JUnit tests
This commit adds unimplemented deleteMax() and max() methods along with
a series of JUnit tests that should indicate the correct functionality
of both methods once implemented. As of now, all of these tests fail.
These methods became necessary when trying to implement the delete(Key
key) method which requires both of these methods.