Add basic DirectedGraph data structure with unimplemented test class
This commit replaces the BinarySearchTree.java and
BinarySearchTreeTest.java files with the DirectedGraph.java and
DirectedGraphTest.java respectively. This change was made after
reviewing the updated assignment specification which required the LCA
algorithm to work on directed acyclic graphs - which could not be
represented in the BinarySearchTree data structure that was previously
implemented. Hence, I've chosen to discard the BinarySearchTree approach
and instead work on a more appropriate data structure. The methods
defined in DirectedGraph.java still require implementations and the test
class requires tests for these methods.