Add multiple LCA implementation for lowestCommonAncestor(int vertex1, int vertex2)
This commit enables the method lowestCommonAncestor(int vertex1, int
vertex2) to return a list of integers as opposed to a single integer, as
was implemented previously. By making this adjustment, along with some
minor modifications, the method can now return all the LCAs of two
vertices. The JUnit tests were updated to reflect this change and all
tests pass.